backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
numpy
tensorflow
pillow
scipy
python-multipart
httpx


